Reduced some code duplication
